Handle missing directories properly
If a directory was not available, Bob was just skipping that line, which
deranged the complete result list. This is now fixed by including the
output to stderr, which results in one line of output for each package,
which is exactly what we want.

Needs current development version ob Bob (v0.16+).