Init: Fix superfluous path separate at end
When no path is provided, then the resulting path has no path separator
at the end.
When a path is provided, the old behaviour was that the resulting path
had a separator at the end.
That is not only inconsistent it also broke the autocommands to open
files in the dist folder as readonly.